Modification: Modified product repository module to follow the proyect structure, separated memory from interface.
