--- docs/api/employees.http
@@ -1,33 +1,0 @@
-### GET request to get all employees
-GET http://localhost:8080/api/v1/employees
-Content-Type: application/json
-
-### GET request to get an specific employee
-GET http://localhost:8080/api/v1/employees/10
-Content-Type: application/json
-
-### POST request to create a new employee
-POST localhost:8080/api/v1/employees
-Content-Type: application/json
-
-{
-  "card_number_id": "98123143",
-  "first_name": "Charles",
-  "last_name": "Santos",
-  "warehouse_id": 1
-}
-
-### PATCH request to update an specific employee
-PATCH localhost:8080/api/v1/employees/3
-Content-Type: application/json
-
-{
-  "card_number_id": "98123143",
-  "first_name": "Charles",
-  "last_name": "Ferreira",
-  "warehouse_id": 1
-}
-
-### DELETE request to delete an specific employee
-DELETE localhost:8080/api/v1/employees/3
-Content-Type: application/json
