--- docs/db/employee.json
@@ -1,16 +1,0 @@
-[
-  {
-    "id": 1,
-    "card_number_id": "402323",
-    "first_name": "Jhon",
-    "last_name": "Doo",
-    "warehouse_id": 1
-  },
-  {
-    "id": 2,
-    "card_number_id": "987323",
-    "first_name": "Paco",
-    "last_name": "Dow",
-    "warehouse_id": 1
-  }
-]
