+++ go.mod
@@ -3,11 +3,9 @@ module github.com/miloalej-dev/W17-G1-Bootcamp
 go 1.24
 
 require (
+	github.com/bootcamp-go/web v1.0.0
 	github.com/go-chi/chi/v5 v5.2.2
 	github.com/go-chi/render v1.0.3
 )
 
-require (
-	github.com/ajg/form v1.5.1 // indirect
-	github.com/bootcamp-go/web v1.0.0 // indirect
-)
+require github.com/ajg/form v1.5.1 // indirect
