+++ internal/application/application_default.go
@@ -3,15 +3,12 @@ package application
 import (
 	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/application/route"
 	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/loader/buyerLoader"
-	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/loader/employee"
 	loaderProduct "github.com/miloalej-dev/W17-G1-Bootcamp/internal/loader/product"
 	loaderWarehouse "github.com/miloalej-dev/W17-G1-Bootcamp/internal/loader/warehouse"
 	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/repository/buyerRepository"
 	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/repository/memory"
-	productRepository "github.com/miloalej-dev/W17-G1-Bootcamp/internal/repository/product"
 	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/service"
 	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/service/buyerService"
-	productService "github.com/miloalej-dev/W17-G1-Bootcamp/internal/service/product"
 	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/service/section"
 	warehouseService "github.com/miloalej-dev/W17-G1-Bootcamp/internal/service/warehouse"
 	"net/http"
@@ -20,6 +17,7 @@ import (
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
+	"github.com/miloalej-dev/W17-G1-Bootcamp/internal/service/product"
 )
 
 // ConfigServerChi is a struct that represents the configuration for ServerChi
@@ -32,8 +30,6 @@ type ConfigServerChi struct {
 	LoaderFilePathProducts string
 	// LoaderFilePath is the path to the file that contains the warehouses
 	LoaderFilePathWarehouse string
-	// LoaderFilePath is the path to the file that contains the warehouses
-	LoaderFilePathEmployee string
 }
 type ServerChi struct {
 	// serverAddress is the address where the server will be listening
@@ -42,7 +38,6 @@ type ServerChi struct {
 	loaderFilePathBuyer     string
 	loaderFilePathProducts  string
 	loaderFilePathWarehouse string
-	loaderFilePathEmployee  string
 }
 
 // NewServerChi is a function that returns a new instance of ServerChi
@@ -66,9 +61,6 @@ func NewServerChi(cfg *ConfigServerChi) *ServerChi {
 		if cfg.LoaderFilePathProducts != "" {
 			defaultConfig.LoaderFilePathProducts = cfg.LoaderFilePathProducts
 		}
-		if cfg.LoaderFilePathEmployee != "" {
-			defaultConfig.LoaderFilePathEmployee = cfg.LoaderFilePathEmployee
-		}
 	}
 
 	return &ServerChi{
@@ -76,7 +68,6 @@ func NewServerChi(cfg *ConfigServerChi) *ServerChi {
 		loaderFilePathBuyer:     defaultConfig.LoaderFilePathBuyer,
 		loaderFilePathProducts:  defaultConfig.LoaderFilePathProducts,
 		loaderFilePathWarehouse: defaultConfig.LoaderFilePathWarehouse,
-		loaderFilePathEmployee:  defaultConfig.LoaderFilePathEmployee,
 	}
 }
 
@@ -94,18 +85,14 @@ func (a *ServerChi) Run() (err error) {
 	ldWarehouse := loaderWarehouse.NewJSONFile(a.loaderFilePathWarehouse)
 	dbWarehouse, err := ldWarehouse.Load()
 
-	ldEmployee := employee.NewJSONFile(a.loaderFilePathEmployee)
-	dbEmployee, err := ldEmployee.Load()
-
 	if err != nil {
 		return
 	}
 	// - repositories
-	rpProduct := productRepository.NewProductMap(dbProduct)
+	rpProduct := memory.NewProductMap(dbProduct)
 	warehouseRepo := memory.NewWarehouseMap(dbWarehouse)
-	sellerRepository := memory.NewSellerMap()
-	employeeRepository := memory.NewEmployeeMap(dbEmployee)
 	rpBuyer := buyerRepository.NewBuyerMap(dbBuyer)
+	sellerRepository := memory.NewSellerMap()
 	sectionRepository := memory.NewSectionMap()
 
 	// - services
@@ -114,14 +101,12 @@ func (a *ServerChi) Run() (err error) {
 	warehouseServ := warehouseService.NewWarehouseDefault(warehouseRepo)
 	sellerService := service.NewSellerService(sellerRepository)
 	sectionService := section.NewSectionDefault(sectionRepository)
-	employeeService := service.NewEmployeeService(employeeRepository)
 
 	// - handlers
 	hdBuyer := handler.NewBuyerHandler(svBuyer)
 	hdProduct := handler.NewProductDefault(svProduct)
 	warehouseHand := handler.NewWarehouseDefault(warehouseServ)
 	sellerHandler := handler.NewSellerHandler(sellerService)
-	employeeHandler := handler.NewEmployeeHandler(employeeService)
 	sectionHandler := handler.NewSectionDefault(sectionService)
 
 	//hd := handler.NewFooHandler()
@@ -136,9 +121,8 @@ func (a *ServerChi) Run() (err error) {
 
 	route.DefaultRoutes(rt)
 	route.BuyerRoutes(rt, hdBuyer)
-	route.WarehouseRoutes(rt, warehouseHand)
 	route.SellerRoutes(rt, sellerHandler)
-	route.EmployeeRoutes(rt, employeeHandler)
+	route.WarehouseRoutes(rt, warehouseHand)
 	route.SectionRoutes(rt, sectionHandler)
 	route.ProductRoutes(rt, hdProduct)
 	/*
