--- pkg/models/employee.go
@@ -1,9 +1,0 @@
-package models
-
-type Employee struct {
-	Id           int    `json:"id"`
-	CardNumberId string `json:"card_number_id"`
-	FirstName    string `json:"first_name"`
-	LastName     string `json:"last_name"`
-	WarehouseId  int    `json:"warehouse_id"`
-}
